production-service-api-platform/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.8
httpx==0.27.2
orjson==3.10.7
prometheus-client==0.20.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
